storage/requirements.txt
connexion[flask]
connexion[uvicorn]
PyYAML
SQLAlchemy
mysqlclient
gunicorn